+++ hypothesis-python/src/hypothesis/strategies/_internal/numbers.py
@@ -50,6 +50,7 @@ Real = Union[int, float, Fraction, Decimal]
 
 class IntegersStrategy(SearchStrategy[int]):
     def __init__(self, start: Optional[int], end: Optional[int]) -> None:
+        super().__init__()
         assert isinstance(start, int) or start is None
         assert isinstance(end, int) or end is None
         assert start is None or end is None or start <= end